api/requirements.txt
flask
psycopg[binary] # PostgreSQL driver
pytest
python-dotenv